statvfs (df) cleanup

+++ lib/fusefs/pathmapper.rb
@@ -12,20 +12,28 @@ module FuseFS
             # @return [Hash<String,MNode>] list of files in a directory, nil for file nodes
             attr_reader :files
 
+            # Useful when mapping a file to store attributes against the
+            # parent directory
             # @return [MNode] parent directory
             attr_reader :parent
 
+            #
             # @return [Hash] metadata for this node
             attr_reader :options
 
+            #
             # @return [String] path to backing file, or nil for directory nodes
             attr_reader :real_path
 
+
             # @!visibility private
-            def initialize(parent_dir)
+            def initialize(parent_dir,stats)
                 @parent = parent_dir
                 @files = {}
                 @options = {}
+                @stats = stats
+                @stats_size = 0
+                @stats.adjust(0,1)
             end
 
             # @!visibility private
@@ -33,6 +41,7 @@ module FuseFS
                 @options.merge!(options)
                 @real_path = real_path
                 @files = nil
+                updated
                 self
             end
 
@@ -76,9 +85,20 @@ module FuseFS
             def[]=(key,value)
                 options[key]=value
             end
+
+            def deleted
+                @stats.adjust(-@stats_size,-1)
+                @stats_size = 0
+            end
+
+            def updated
+                new_size = File.size(real_path)
+                @stats.adjust(new_size - @stats_size)
+                @stats_size = new_size
+            end
         end
 
-        # Convert FuseFS raw_mode strings to IO open mode strings
+        # Convert FuseFS raw_mode strings back to IO open mode strings
         def self.open_mode(raw_mode)
             case raw_mode
             when "r"
@@ -106,6 +126,10 @@ module FuseFS
         #     default is false
         attr_accessor  :allow_write
 
+        #
+        # @return [StatsHelper] accumulated filesystem statistics
+        attr_reader :stats
+        
         # Creates a new Path Mapper filesystem over an existing directory
         # @param [String] dir
         # @param [Hash] options
@@ -114,8 +138,8 @@ module FuseFS
         # @see #initialize
         # @see #map_directory
         def PathMapperFS.create(dir,options={ },&block)
-            pm_fs = PathMapperFS.new(options)
-            pm_fs.map_directory(dir,&block) 
+            pm_fs = self.new(options)
+            pm_fs.map_directory(dir,&block)
             return pm_fs
         end
 
@@ -123,8 +147,13 @@ module FuseFS
         # @param [Hash]  options
         # @option options [Boolean] :use_raw_file_access
         # @option options [Boolean] :allow_write
+        # @option options [Integer] :max_space available space for writes (for df)
+        # @option options [Integer] :max_nodes available nodes for writes (for df)
         def initialize(options = { })
-            @root = MNode.new(nil)
+            @stats = StatsHelper.new()
+            @stats.max_space = options[:max_space]
+            @stats.max_nodes = options[:max_nodes]
+            @root = MNode.new(nil,@stats)
             @use_raw_file_access = options[:use_raw_file_access]
             @allow_write = options[:allow_write]
         end
@@ -230,9 +259,9 @@ module FuseFS
 
         # @!visibility private
         def write_to(path,contents)
-            File.open(unmap(path),"w") do |f|
-                f.print(contents)
-            end
+            node = node(path)
+            File.open(node.real_path,"w") { |f| f.print(contents) }
+            node.updated
         end
 
         # @!visibility private
@@ -300,33 +329,61 @@ module FuseFS
             file.sysseek(offset)
             file.syswrite(buf[0,sz])
         end
+        
+        # @!visibility private
+        def raw_sync(path,datasync,file=nil)
+            file = @openfiles[path] unless file
+            if datasync
+                file.fdatasync
+            else
+                file.sync
+            end
+        end
 
         # @!visibility private
         def raw_close(path,file=nil)
-            unless file
-                file = @openfiles.delete(path)
+            file = @openfiles.delete(path) unless file
+           
+            if file && !file.closed?
+                begin
+                    flags = file.fcntl(Fcntl::F_GETFL) & Fcntl::O_ACCMODE
+                    if flags == Fcntl::O_WRONLY || flags == Fcntl::O_RDWR
+                        #update stats
+                        node = node(path)
+                        node.updated if node
+                    end
+                ensure
+                    file.close
+                end
             end
-            file.close if file
+
+        end
+
+        # @!visibility private
+        def statistics(path)
+            @stats.to_statistics
         end
 
         private
        
         def make_node(path)
-            #split path into components 
+            #split path into components
             components = path.to_s.scan(/[^\/]+/)
             components.inject(@root) { |parent_dir, file|
-                parent_dir.files[file] ||= MNode.new(parent_dir)
+                parent_dir.files[file] ||= MNode.new(parent_dir,@stats)
             }
         end
         
         def recursive_cleanup(dir_node,&block)
-            dir_node.files.delete_if do |path,child| 
-                if child.file?
+            dir_node.files.delete_if do |path,child|
+                del = if child.file?
                     yield child
                 else
-                    recursive_cleanup(child,&block) 
+                    recursive_cleanup(child,&block)
                     child.files.size == 0
                 end
+                child.deleted if del
+                del
             end
         end
     end
